+++ tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(downloadR)
+
+test_check("downloadR")