+++ NEWS.md
@@ -1,3 +1,4 @@
 # downloadR 0.99.0
 
 * Added a `NEWS.md` file to track changes to the package.
+* Added axel, wget, and download.file as methods.